pyproject.toml
[tool.poetry]
name = "telegram-stars-starter"
version = "0.1.0"
description = "Telegram stars starter template"
authors = ["fauzaanu <fauzaangasim@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^21.3"
python-dotenv = "^1.0.1"
openai = "^1.33.0"
peewee = "^3.17.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
